Refactor event handling and improve type safety
- Added default return cases in `EventNameToString` and `EventTypeTypeToString`.
- Removed unnecessary `constexpr` key constants in the `KEY` enum.
- Updated `Logger` class method signature for clarity.
- Reorganized includes in `Application.h`.
- Deleted `Square1.mesh` and noted differences in `Scene.cn`.
- Modified `Mesh` class to use `size_t` for vertex and index sizes.
- Updated input handling in `Window.cpp` to use `reinterpret_cast`.
- Changed FPS handling in `Panel` and `ViewportPanel` from `float` to `int`.
- Updated `Physics` class to use `float` for `mDeltaTime`.
- Enhanced read/write methods in `Reader` and `Writer` to use `size_t`.
- Added binary data to `Square.mesh`.

+++ Engine/src/Editor/Editor.h
@@ -1,9 +1,9 @@
 #pragma once
-#include <GL/glew.h>
-#include <GLFW/glfw3.h>
 #include "../EventSystem/Subject/Subject.h"
 #include "Panel/Panel.h"
 #include "../Locator/Locator.h"
+#include <GL/glew.h>
+#include <GLFW/glfw3.h>
 class Window;
 class Editor : public Observer, public Subject
 {
@@ -13,7 +13,7 @@ class Editor : public Observer, public Subject
 public:
 	Editor();
 	void Init();
-	void Draw(float fps);
+	void Draw(int fps);
 	void OnClose();
 	using Subject::AddObserver;
 };
